pyproject.toml
[tool.poetry]
name = "jarvis-auth"
version = "0.1.0"
description = "FastAPI authentication service for Jarvis."
authors = ["Jarvis Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.104.1"
uvicorn = { version = ">=0.24.0", extras = ["standard"] }
sqlalchemy = ">=2.0.23"
alembic = ">=1.12.1"
psycopg2-binary = ">=2.9.9"
passlib = { version = ">=1.7.4", extras = ["bcrypt"] }
bcrypt = "4.1.3"
python-jose = { version = ">=3.3.0", extras = ["cryptography"] }
pydantic = ">=2.5.2"
pydantic-settings = ">=2.1.0"
python-multipart = ">=0.0.6"
httpx = ">=0.25.2"
email-validator = ">=2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.3"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

